requirements.txt
lxml==4.9.1
Pillow==9.3.0
python-pptx==0.6.21
XlsxWriter==3.0.3
